browser-switch/src/test/java/com/braintreepayments/browserswitch/BrowserSwitchRequestTest.java
package com.braintreepayments.browserswitch;

import org.json.JSONException;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.RobolectricTestRunner;

import static org.junit.Assert.assertEquals;

@RunWith(RobolectricTestRunner.class)
public class BrowserSwitchRequestTest {

    @Test
    public void fromJson() throws JSONException {

        String json = "{\n" +
                "  \"requestCode\": 123,\n" +
                "  \"url\": \"https://example.com\",\n" +
                "  \"state\": \"PENDING\"\n" +
                "}";
        BrowserSwitchRequest sut = BrowserSwitchRequest.fromJson(json);

        assertEquals(sut.getRequestCode(), 123);
        assertEquals(sut.getUri().toString(), "https://example.com");
        assertEquals(sut.getState(), BrowserSwitchRequest.PENDING);
    }

    @Test
    public void toJson() throws JSONException {
        String json = "{\n" +
                "  \"requestCode\": 123,\n" +
                "  \"url\": \"https://example.com\",\n" +
                "  \"state\": \"SUCCESS\"\n" +
                "}";
        BrowserSwitchRequest original = BrowserSwitchRequest.fromJson(json);
        BrowserSwitchRequest restored = BrowserSwitchRequest.fromJson(original.toJson());

        assertEquals(restored.getRequestCode(), original.getRequestCode());
        assertEquals(restored.getUri(), original.getUri());
        assertEquals(restored.getState(), original.getState());
    }
}
